feat(): export `EmptyResult` and `EmptyResultStatus` separately
for better tree-shaking for assets

BREAKING CHANGE: `EmptyResult` and `EmptyResultStatus` are exported separately now
