+++ libs/basic-components/EmptyResult.js
@@ -1,0 +1,1 @@
+export * from "./dist/esm/EmptyResult";
