+++ libs/basic-components/src/EmptyResultStatus.ts
@@ -1,0 +1,9 @@
+export enum EmptyResultStatus {
+  BrowserTooOld = "browser-too-old",
+  Empty = "empty",
+  NoData = "no-data",
+  NoHistoryVersion = "no-history-version",
+  NoVisitRecord = "no-visit-record",
+  SearchEmpty = "search-empty",
+  WelcomeToCreate = "welcome-to-create",
+}
