+++ libs/basic-components/src/index.ts
@@ -3,7 +3,6 @@ export * from "./Link";
 export * from "./HashLink";
 export * from "./Sidebar";
 export * from "./GeneralIcon";
-export * from "./EmptyResult";
 export * from "./ItemActionsComponent";
 export * from "./utils";
 export * from "./ZoomPanel/ZoomPanel";
